+++ unnaf/src/platform.h
@@ -12,6 +12,7 @@
 #include <string.h>
 #include <stdio.h>
 #include <stdarg.h>
+#include <ctype.h>
 #include <unistd.h>
 #include <sys/stat.h>